Largest improvements and bug fixes
* Code mobility: improvements of garbage collection of dynamically
loaded classes.

* Added remote pointer routing for pointers that originate from a remote
RmiRegistry that does not have the RMI server listener enabled.

* Bug fix on evaluation of equals() method in Skeleton.invoke().

* Bug fix on remote pointer sharing (the shared port of the pointer is
the one of the remote RMI listener).

* Bug fix on garbage collection of new unnamed skeletons that have never
been referenced (due to connection latency or errors, for instance).

* Updated License headers in all files.

* Updated the handshake protocol in RmiHandler (now also the ports of
the listeners are exchanged)


Next TODO:

* Dynamic loaded class should be associated with the authidentifier that
requested their loading. So an authIdentifier can be notified, when it
tries to load a class already dynamically loaded by another.