Distributed Garbage Collection and Remote Invocation improvements
* Added Garbage Collection mechanism based on lease time.

* Added possibility to switch to a plain old RPC (Remote Procedure Call)
mode, disabling automatic referencing of remote objects and disabling
the Distributed Garbage Collector service. Only objects that are
published on registry can be accessible from remote nodes in this mode.

* Added the @RMIAsync annotation, used to mark a void interface method
as asynchronous. This can be useful to continue execution without
waiting for RMI response on void methods only.

* Added @RMICached annotation, used to mark a no-arguments method to
cache its return value for a specified amount of time. For instance, it
is useful to reduce RMI communication at the RMI level when the
application calls frequently a method whose return value changes slow in
time.

* Various minor bug fixes